preload/dynamic_removing_preload.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure that dynamically removed preloaded resource stop downloading');
</script>
<body>
<script>
    var link = document.createElement("link");
    link.as = "script";
    link.rel = "preload";
    link.href = "resources/dummy.js";
    document.body.appendChild(link);
    document.body.removeChild(link);
    window.addEventListener("load", t.step_func(function() {
        assert_equals(performance.getEntriesByType("resource").length, 3);
        t.done();
    }));
</script>
<script src="resources/dummy.js?pipe=trickle(d0.2)"></script>
</body>
